add currency

+++ src/main/java/hu/kits/investments/domain/marketdata/AssetPrices.java
@@ -1,5 +1,6 @@
 package hu.kits.investments.domain.marketdata;
 
+import java.math.BigDecimal;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
@@ -8,13 +9,13 @@ import hu.kits.investments.domain.asset.Asset;
 
 public class AssetPrices {
 
-    private final Map<Asset, Double> priceMap;
+    private final Map<Asset, BigDecimal> priceMap;
 
-    public AssetPrices(Map<Asset, Double> priceMap) {
+    public AssetPrices(Map<Asset, BigDecimal> priceMap) {
         this.priceMap = priceMap;
     }
     
-    public Optional<Double> price(Asset asset) {
+    public Optional<BigDecimal> price(Asset asset) {
         return Optional.ofNullable(priceMap.get(asset));
     }
