+++ src/main/java/hu/kits/investments/domain/marketdata/PriceData.java
@@ -1,7 +1,8 @@
 package hu.kits.investments.domain.marketdata;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
-public record PriceData(String ticker, LocalDate date, double price) {
+public record PriceData(String ticker, LocalDate date, BigDecimal price) {
 
 }
